third_party/blink/web_tests/platform/generic/fast/multicol/newmulticol/balance3-expected.txt

PASS mc.offsetHeight is ref.offsetHeight
PASS successfullyParsed is true

TEST COMPLETE


There should be two identical boxes below.

line

line
